ImageName::from_path
